Make supervision tree fetching more error proof
If a child is wrongly specified as `supervisor`, but it does not
implement this behaviour, the `supervisor:which_children/0` call can
return unpredictable values. (Especially if the worker is a `gen_server`
which can return any sort of funny value in its `handle_call`).

This issue has been seen in more than one open-source library.

Hopefully closes #37